+++ styx/src/engine/scene/CFrustum.hpp
@@ -2,6 +2,8 @@
 
 #include <array>
 
+#include "src/engine/helper/Types.hpp"
+
 #include "src/engine/helper/geom/CPlane.hpp"
 
 class CFrustum
@@ -9,7 +11,7 @@ class CFrustum
 public:
 	explicit CFrustum( const glm::mat4 &viewProjectionMatrix );
 
-	bool IsSphereInside( const glm::vec3 &position, const float sphereRadius ) const;
+	bool IsSphereInside( const glm::vec3 &position, const f16 sphereRadius ) const;
 
 private:
 	std::array< CPlane, 6 > m_planes;
